refactor: 카드 등록 컴포넌트 폴더 정리
