+++ src/components/CardRegister/CardCompanySelector/CardCompanySelector.tsx
@@ -5,12 +5,12 @@ import { useEffect, useRef, useState } from "react";
 import { CARD_COMPANIES } from "@/constants/cardCompanies";
 import type { CardCompany } from "@/constants/cardCompanies";
 import FormField from "@components/common/FormField";
-import type { AddCardFormStepKey } from "@/constants/cardForm";
+import type { CardRegisterFormStepKey } from "@/constants/cardForm";
 
 interface CardCompanySelectorProps {
   cardCompany: CardCompany | null;
   onSelect: (cardCompany: CardCompany) => void;
-  onNextStep: (currentStepKey: AddCardFormStepKey) => void;
+  onNextStep: (currentStepKey: CardRegisterFormStepKey) => void;
 }
 
 function CardCompanySelector({
