+++ src/components/CardRegister/CardPasswordInputField/constants.ts
