+++ src/stories/CardCVCInputField.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { userEvent, within } from "storybook/test";
 
-import CardCVCInputField from "@components/CardCVCInputField/CardCVCInputField";
+import CardCVCInputField from "@/components/CardRegister/CardCVCInputField/CardCVCInputField";
 
 const meta = {
   title: "CardCVCInputField",
