+++ src/stories/CardCompanySelector.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { userEvent, within } from "storybook/test";
 
-import CardCompanySelector from "@components/CardCompanySelector/CardCompanySelector";
+import CardCompanySelector from "@/components/CardRegister/CardCompanySelector/CardCompanySelector";
 import { CARD_COMPANIES } from "@/constants/cardCompanies";
 
 const meta = {
