+++ src/stories/CardNumberInputField.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { userEvent, within } from "storybook/test";
 
-import CardNumberInputField from "@components/CardNumberInputField/CardNumberInputField";
+import CardNumberInputField from "@/components/CardRegister/CardNumberInputField/CardNumberInputField";
 
 const meta = {
   title: "CardNumberInputField",
