+++ src/stories/CardValidityPeriodInputField.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { userEvent, within } from "storybook/test";
 
-import CardValidityPeriodInputField from "@components/CardValidityPeriodInputField/CardValidityPeriodInputField";
+import CardValidityPeriodInputField from "@/components/CardRegister/CardValidityPeriodInputField/CardValidityPeriodInputField";
 
 const meta = {
   title: "CardValidityPeriodInputField",
