+++ src/utils/validator.ts
@@ -1,7 +1,7 @@
-import type { CardNumberUnits } from "@/components/CardNumberInputField/CardNumberInputField";
+import type { CardNumberUnits } from "@/components/CardRegister/CardNumberInputField/CardNumberInputField";
 import { detectCardBrand, getCardNumberFormat } from "./card";
 import type { CardCompany } from "@/constants/cardCompanies";
-import type { ValidityPeriod } from "@/components/CardValidityPeriodInputField/CardValidityPeriodInputField";
+import type { ValidityPeriod } from "@/components/CardRegister/CardValidityPeriodInputField/CardValidityPeriodInputField";
 import { CARD_FIELD_LENGTH } from "@/constants/cardForm";
 
 export const checkIsOnlyDigits = (input: string) => /^\d*$/.test(input);
